CHROMIUM: cpufreq: keep freezer happy when not current governor
Fix a problem where the hung task mechanism was deeming the interactive
clock boost thread as hung.  This was because the thread is created at
module init but never run/woken up until needed.  If the governor is not
being used this can be forever.  To workaround this explicitly wake up
the thread once all the necessary data structures are initialized.  The
latter required some minor code shuffle.

Signed-off-by: Sam Leffler <sleffler@chromium.org>

BUG=none
TEST=boot with default governor ondemand; leave the system idle for >120 secs (hung_timeout_task_secs)

Change-Id: Ie2c058dd75dcb6460ea10e7ac997e46baf66b1fe
Reviewed-on: https://gerrit.chromium.org/gerrit/26262
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Ready: Sam Leffler <sleffler@chromium.org>