CHROMIUM: ALSA: hda_generic - Add pin switch controls.
If the auto_mute functionality is suppressed by a hint then expose
the pin switches.  This way user space can control the muting behavior
based on the jack states that are reported to it through the hda_jack
interface.  Before the only options were automatic or no control.

With auto-mic there is already a 'Capture Mux' exported when the
automatic switch is disabled.

BUG=chromium:260974
TEST=use amixer to set new controls, check that they mute.

Change-Id: I7570db119ba75dccf2d3d59309c15e6769124b27
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62305
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>