CHROMIUM: ALSA: usb-audio: add quirk for Jabra speakerphone
The Jabra speakerphones have buttons which can change directly the value
of the volume or the muted state.
If the driver is caching those control values, it is no longer properly
sync with the device side after button presses have happened, so we add
a quirk for those devices to avoid caching control values.

Also fix the volume min/max/resolution as the device is sending back
erroneous values.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chromium:278394
TEST=connect a Jabra Speakerphone 410, check volumes and mute with
amixer, change the speaker volume and the mic mute state with the
speakerphone button, re-run amixer and see updated values

Change-Id: Icaafd12d65ccb55a48a0225bfc7b1362d8f7986e
Reviewed-on: https://chromium-review.googlesource.com/168405
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>

[benzh: 3.14 rebase. Resolved trivial conflicts]
Signed-off-by: Ben Zhang <benzh@chromium.org>

Conflicts:
	sound/usb/mixer.c
[filbranden: 3.18 rebase. Solved trivial whitespace conflict.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>