CHROMIUM: cpufreq: interactive: don't defer input open to workqueue
cpufreq_interactive defers opening of input devices from the
->connect() callback to a workqueue.  This introduces races
with the ->disconnect() callback and within the input core as
it expects the input_open_device() call to happen in ->connect().
The only advantage to deferring to a workqueue would be to improve
CPU up/down latency, but this should already be a low overhead
operation on systems where the UI is already monitoring input
devices.

BUG=None
TEST=Rapidly hotplug CPUs and a USB mouse; no crashes observed.

Change-Id: Ic5ffd117538a1b0349c07324f79aca7b76794f38
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186243
Reviewed-by: Dylan Reid <dgreid@chromium.org>