UPSTREAM: ASoC: rt5677: Support DSP function for VAD application
The ALC5677 has a programmable DSP, and there is a SPI that is dadicated for DSP
firmware loading. Therefore, the patch includes a SPI driver for writing the DSP
firmware. The VAD(Voice Activaty Detection) has be implemented and use the DSP to
recognize the key phase.

Signed-off-by: Oder Chiou <oder_chiou@realtek.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

Conflicts:
	sound/soc/codecs/Makefile
	sound/soc/codecs/rt5677.c

Change-Id: Ia8a5ddea16c73a14da52777bf2fa24830644e900
Reviewed-on: https://chromium-review.googlesource.com/224141
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Anatol Pomazau <anatol@google.com>
Commit-Queue: Anatol Pomazau <anatol@google.com>
Tested-by: Anatol Pomazau <anatol@google.com>

Conflicts:
	sound/soc/codecs/Makefile
	sound/soc/codecs/rt5677.c
[filbranden: 3.18 rebase. Solved trivial conflicts.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>