CHROMIUM: ASOC: max98090: Shutdown codec before updating some registers
According to user manual, there are some registers whose change requires
shutdown bit = 0. This patch adds shutdown call before updating those
registers. The codec state will be restored after the update.

In max98090_dai_hw_params, codec might get new lrclk and pcm format.
If pcm format is changed, or needs to be set for the first time,
it needs a shutdown.
If lrclk is changed, or needs to be set for the first time, it needs a
shutdown.
If sysclk is changed, it needs a shutdown.

Apart from the registers listed on user manual, DMIC should be toggled
when codec is off. However, in reality, it is fine to enable DMIC when
codec is on, as long as there is an extra shutdown/restore after DMIC is
enabled.

This patch adds an event callback to DAPM supply widget DMICL_ENA and
DMICR_ENA so DAPM will shutdown codec before DMIC is enabled.
After DMIC is enabled, it does another shutdown/restore since shutdown
register is not protected in the time frame between PRE_PMU and POST_PMU
events.
Note that this is only required for max98090, not max98091.

In methods other than dmic_event, the code to shutdown codec, update
registers, and turn on codec are protected by a mutex in private data to
make sure codec is in shutdown state when register is being updated.

With this patch, there will be no audio drop in playback stream when
capture stream starts to use external mic. There will be audio drop in
playback stream when capture stream starts to use digital mic.

To avoid the playback stream audio drop, dmic can be enabled in all
playback cases, which will be handled in machine driver by adding DAPM
routes.

The patch (https://chromium-review.googlesource.com/#/c/233863/9) on
3.10 kernel, apply this patch to 3.14 kernel.

BUG=chrome-os-partner:35066
BUG=chromium:426595
TEST=on jerry2 board: play youtube at the same time, record audio using
recording App(Mic Note), then switch to/from external/internal Mic.
Check recorded audio is fine.

Change-Id: Ic1838f9f1c7e0da779c5dfcb907ae6933688e838
Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/238325
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>

Conflicts:
	sound/soc/codecs/max98090.c
[filbranden: 3.18 rebase. Solved MAJOR conflicts.]
[TODO(filbranden): REALLY ask sonnyrao@ to check this one.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>