CHROMIUM: ASoC: rt5677: Refactor SPI driver
Added rt5677_spi_read() and refactored rt5677_spi_write()
so that an arbitrary block in the DSP address space can
be read/written via SPI.

An ACPI match table is added so that the driver can
be instantiated for ACPI device ID 'RT5677AA'. The ID is
temporary and we should switch to an official ID once
it's assigned.

BUG=chrome-os-partner:33495
TEST=load firmware via SPI; hotword detection works
CQ-DEPEND=CL:240280

Change-Id: I97f0405c33bee6f23fe7faeabddd29472a8ee6ea
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/240209
Reviewed-by: Dylan Reid <dgreid@chromium.org>