FROMLIST: mfd: syscon: Add register stride to DT bindings.
This patch adds register stride to dt bindings so that the consumers of
the syscon could change it to their need. One of the the use cases for
this feature is Qualcomm qfprom which needs a byte access to regmap
returned from syscon.
(from https://patchwork.kernel.org/patch/5707771/)

BUG=None
TEST=build/boot on storm succeded

Change-Id: I067040d33f54f4b08594e296b7d9dd8420eda942
Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Signed-off-by: Narendran Rajan <nrajan@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/246450
Tested-by: Narendran M R <narendranmr@gmail.com>
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Commit-Queue: Grant Grundler <grundler@chromium.org>