CHROMIUM: pstore/ram: verify ramoops header before saving record
On some devices the persistent memory contains junk after a cold boot,
and /dev/pstore/dmesg-ramoops-* are created with random data which is
not the result of a kernel crash.

This patch adds a ramoops header check and skips any
persistent_ram_zone that does not have a valid header.

BUG=chromium:443764
TEST=platform_KernelErrorPaths, cold boot, corrupted header on samus

Change-Id: I668ed186a6d35568d497e989d6e939f5fcf1c53a
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/244157
Reviewed-by: Kees Cook <keescook@chromium.org>